requirements.txt
ollama 
openai 
anthropic 
loguru
langfuse
python-dotenv
rich
questionary
pyyaml